requirements.txt
pandas
numpy<2
openpyxl
pyyaml
xgboost
scikit-learn
typer
pytest
streamlit
